refactor: move AasServiceOperation to aasservice
