+++ basyx.aasrepository/basyx.aasrepository-backend/src/main/java/org/eclipse/digitaltwin/basyx/aasrepository/backend/AasRepositoryBackend.java
@@ -1,6 +1,7 @@
 package org.eclipse.digitaltwin.basyx.aasrepository.backend;
 
 import org.eclipse.digitaltwin.aas4j.v3.model.AssetAdministrationShell;
+import org.eclipse.digitaltwin.basyx.aasservice.AasServiceOperations;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
